+++ tests/unit/test_tfs.py
@@ -57,6 +57,8 @@ def sagemaker_session():
         boto_region_name=REGION,
         config=None,
         local_mode=False,
+        s3_resource=None,
+        s3_client=None,
     )
     session.default_bucket = Mock(name="default_bucket", return_value="my_bucket")
     session.expand_role = Mock(name="expand_role", return_value=ROLE)